refactor(adr-061): split clang template-specialization parsing into extract/headers
ADR-061 Phase 5 item 1's last entity module -- closes the CastXML/Clang
parser split IN FULL on both backends.

castxml has no separate template-entity module. Investigated
dumper_castxml.py and every module in extract/headers/castxml/ for
"template"/"specialization"/"Specialization" and found nothing: castxml's
XML output resolves a class-template specialization down to an ordinary
Struct/Class element, indistinguishable at the AST-node level from a
non-template record -- no ClassTemplateSpecializationDecl-shaped node, no
separate specialization-index pass, no RecordType.is_template_pattern
concept at all. Every fact a specialization's own record carries (fields,
bases, vtable/RTTI layout) is already produced by the ordinary record path
records.py owns. This is a documented "nothing here" finding, not a
deferred one -- dumper_castxml.py's and context.py's module docstrings now
say so explicitly.

extract/headers/clang/templates.py holds what clang genuinely needed:
template-parameter-kind/default/name reconstruction and specialization-
spelling/indexing (_index_template_param_kinds/_defaults/_names,
_specialization_spelling, build_specialization_index), moved out of
dumper_clang_vtable.py as free functions, unchanged in body --
dumper_clang_vtable.py (despite its name, always held two loosely related
halves: record/vtable layout, which stayed, and this template half, which
didn't) re-exports every name for existing direct importers.
dumper_clang_vtable.py shrinks from 1273 lines to 638. No parse_templates()
entry point exists parallel to parse_enums/parse_functions/parse_types: a
ClassTemplateSpecializationDecl never joins one of _ClangAstParser's own
categorized _Decl lists, so _walk/_categorize stayed in dumper_clang.py.

Two dependency-direction findings fixed proactively (per the "public-ize
in place, check both import directions before physically moving" lesson
the prior four slices' Codex reviews established):
- _SCOPE_NODE_KINDS's canonical definition moved from dumper_clang_expr.py
  (which extract may not import -- it pulls in diff_cxx_rules, classified
  compare) into templates.py itself, with dumper_clang_expr.py reading it
  back.
- build_specialization_index takes is_record_definition as an explicit,
  required keyword-only parameter rather than importing it from
  dumper_clang_vtable.py, since that module re-exports every templates.py
  name for back-compat -- importing is_record_definition back would make
  the two modules import each other.

Zero output/snapshot change -- every migrated name is a one-line
delegation, and every existing caller (including tests reading private
attributes/methods directly) is unaffected.

Verified: check_architecture.py (0 errors, including the debt-no-growth
gate and abicheck/extract/AGENTS.md's 150-line package-instructions cap),
mypy (clean), ruff check (clean on changed files -- ruff format --check
flags only pre-existing, unrelated formatter-version drift on code moved
verbatim from dumper_clang_vtable.py), the castxml/clang/dumper/template/
specialization-focused suite (3471 passed), and real clang integration
tests covering template specializations (5 passed).
