+++ changelog.d/20260829_090342_noreply_adr_061_remaining_items_i93kmd.md
@@ -1,0 +1,3 @@
+### Changed
+
+- **Clang template-specialization parsing moved into `extract/headers/clang/templates.py`** — ADR-061 Phase 5 item 1's last entity module, closing out the CastXML/Clang parser split in full. The template-parameter-kind/default/name reconstruction and specialization-spelling/indexing machinery (`_index_template_param_kinds`/`_index_template_param_defaults`/`_index_template_param_names`, `_specialization_spelling`, `build_specialization_index`) moved out of `dumper_clang_vtable.py` as free functions, unchanged in behavior; `dumper_clang_vtable.py` re-exports every name for existing direct importers. CastXML has no separate template-entity module — its XML output resolves a class-template specialization down to an ordinary record node, so there was nothing to split out there; `dumper_castxml.py`'s module docstring now documents this explicitly. Pure internal refactor: zero output/snapshot change.
